feat(Tabs): change getters to optional parametrs
closes #2220
