+++ src/components/Tabs/Tab/TabsTab.tsx
@@ -2,31 +2,20 @@ import './TabsTab.css';
 
 import React, { forwardRef } from 'react';
 
-import { IconComponent, IconPropSize } from '../../../icons/Icon/Icon';
+import { IconPropSize } from '../../../icons/Icon/Icon';
 import { cnMixFocus } from '../../../mixs/MixFocus/MixFocus';
 import { cn } from '../../../utils/bem';
 import { getByMap } from '../../../utils/getByMap';
-import { TabsPropSize } from '../Tabs';
+import { TabsPropSize, TabsTabProps } from '../types';
 
 export const cnTabsTab = cn('TabsTab');
-
-type Props = {
-  size: TabsPropSize;
-  onlyIcon?: boolean;
-  icon?: IconComponent;
-  iconSize?: IconPropSize;
-  label: string;
-  checked: boolean;
-  onChange: React.MouseEventHandler<HTMLButtonElement>;
-  className?: string;
-};
 
 const sizeMap: Record<TabsPropSize, IconPropSize> = {
   s: 'xs',
   m: 's',
 };
 
-export const TabsTab = forwardRef<HTMLButtonElement, Props>((props, ref) => {
+export const TabsTab = forwardRef<HTMLButtonElement, TabsTabProps>((props, ref) => {
   const {
     label,
     onChange,
