+++ src/components/Tabs/__stories__/examples/TabsExampleIcon/TabsExampleIcon.tsx
@@ -10,21 +10,21 @@ import { Tabs } from '../../../Tabs';
 
 type Item = {
   name: string;
-  icon?: IconComponent;
+  image?: IconComponent;
 };
 
 const items: Item[] = [
   {
     name: 'Первый',
-    icon: IconPhoto,
+    image: IconPhoto,
   },
   {
     name: 'Второй',
-    icon: IconRing,
+    image: IconRing,
   },
   {
     name: 'Третий вариант',
-    icon: IconCamera,
+    image: IconCamera,
   },
 ];
 
@@ -37,14 +37,14 @@ export const TabsExampleIcon = () => {
         onChange={({ value }) => setValue(value)}
         items={items}
         getLabel={(item) => item.name}
-        getIcon={(item) => item.icon}
+        getIcon={(item) => item.image}
       />
       <Tabs
         value={value}
         onChange={({ value }) => setValue(value)}
         items={items}
         getLabel={(item) => item.name}
-        getIcon={(item) => item.icon}
+        getIcon={(item) => item.image}
         onlyIcon
       />
     </StoryBookExample>
